feat!(nix): using nix as package and dependency manager
Nix is a package manager --such as apt-- based on nix, a functional programming language. We can use nix to declare global dependencies of this project -- such as pnpm and biome --; as well as project-specific.

The main reason I'm looking to use nix in this repo is I'll be able to make reproducible development environtments, declare and pin dependencies, to ensure a local-first CI pipeline. We can even containerized such workflow, allowing is to deploy, almost, everywhere.

This is the first step of our nix integration. In this commit, there's a `shell.nix` file that deploys our development environtment. It currently pins to nixpkgs repository that you configured when you installed NixOS or Nix.

This file was generated using `lorri` (https://github.com/nix-community/lorri/) and we intend to deploy the next toolkit:

lorri, for shell.nix and related files management,
direnv, for automatic integration to shell.nix -- and .envrc / .env --
npins, to pin specific dependency, a-la package-lock.json