chore: magick jpg to webp
Converted jpg to webp.

We should write a script to do this on a regular basis. I would say a `find public/ -type f -exec file --mime-type {} \;` with some grep or awk processing to filter in every image mime-type and filter out webp and svg.

That should get us covered.