Add FLISOL Guayaquil 2025 event to events list (#74)
* Add FLISOL Guayaquil 2025 event to events list

* Update FLISOL event info and add new event image

Updates the FLISOL event information for 2025 and adds a
relevant image.

* chore: magick jpg to webp

Converted jpg to webp.

We should write a script to do this on a regular basis. I would say a `find public/ -type f -exec file --mime-type {} \;` with some grep or awk processing to filter in every image mime-type and filter out webp and svg.

That should get us covered.

* chore: change related filepath

---------

Co-authored-by: Brayen Gavilanes <bjgavilanes@gmail.com>